requirements.txt
google-cloud-aiplatform~=1.50.0
kfp~=2.7.0
google-cloud-storage~=2.16.0

# Machine Learning and Data Handling Libraries
pandas~=2.2.2
scikit-learn~=1.4.2
numpy~=1.26.4
joblib~=1.4.2